+++ GUIDS.cs
@@ -45,7 +45,8 @@ namespace TrailsPlugin {
         public static readonly Guid FilteredStatisticsFilter = new Guid("92750940-2359-11e0-ac64-0800200c9a66");
 
         //Generated Trails
-        public static readonly Guid SplitsTrail    = new Guid("30681520-b220-11e0-a00b-0800200c9a66");
+        public static readonly Guid SplitsTrail = new Guid("30681520-b220-11e0-a00b-0800200c9a66");
+        public static readonly Guid SwimSplitsTrail = new Guid("c35b7cad-cb7a-4040-a498-11ad813aa0ef");
         public static readonly Guid ReferenceTrail = new Guid("30681521-b220-11e0-a00b-0800200c9a66");
         public static readonly Guid HighScoreTrail = new Guid("30681522-b220-11e0-a00b-0800200c9a66");
         public static readonly Guid UniqueRoutesTrail = new Guid("389f4602-6480-4e7c-9830-3ec04a48aa6d");